feat: Added property IsSuccess in Results
